Prepare to sync with main

+++ zs_train.py
@@ -13,13 +13,14 @@
 # limitations under the License.
 
 import os
+import sys
 
 import torch
 import torch.optim as optim
 from torch import nn
 
 from config import cfg
-from models import lenetf, resnetf, vggf
+from models import default_model_path, init_models_faulty
 
 # import torch.nn.functional as F
 # import torchvision
@@ -30,46 +31,22 @@ debug = False
 torch.manual_seed(0)
 
 
-def init_models(arch, precision, retrain, checkpoint_path):
-
-    in_channels = 3
-
-    """
-    unperturbed model
-    """
-    if arch == "vgg11":
-        model = vggf("A", in_channels, 10, True, precision, 0, 0, 0, 0, [])
-    elif arch == "vgg16":
-        model = vggf("D", in_channels, 10, True, precision, 0, 0, 0, 0, [])
-    elif arch == "resnet18":
-        model = resnetf("resnet18", 10, precision, 0, 0, 0, 0, [])
-    elif arch == "resnet34":
-        model = resnetf("resnet34", 10, precision, 0, 0, 0, 0, [])
-    else:
-        model = lenetf(in_channels, 10, precision, 0, 0, 0, 0, [])
-
-    print(model)
-
-    checkpoint_epoch = 0
-    if retrain:
-        print("Restoring model from checkpoint", checkpoint_path)
-        checkpoint = torch.load(checkpoint_path)
-
-        model.load_state_dict(checkpoint["model_state_dict"])
-        print("restored checkpoint at epoch - ", checkpoint["epoch"])
-        print("Training loss =", checkpoint["loss"])
-        print("Training accuracy =", checkpoint["accuracy"])
-        checkpoint_epoch = checkpoint["epoch"]
-
-    return model, checkpoint_epoch
-
-
 def training(
-    trainloader, arch, dataset, precision, retrain, checkpoint_path, device
+    trainloader,
+    arch,
+    dataset,
+    precision,
+    retrain,
+    checkpoint_path,
+    force,
+    device,
+    fl,
+    ber,
+    pos,
 ):
 
-    model, checkpoint_epoch = init_models(
-        arch, precision, retrain, checkpoint_path
+    model, checkpoint_epoch = init_models_faulty(
+        arch, precision, retrain, checkpoint_path, fl, ber, pos
     )
 
     print("Training with Learning rate %.4f" % (cfg.learning_rate))
@@ -80,7 +57,9 @@ def training(
     torch.backends.cudnn.benchmark = True
 
     curr_lr = cfg.learning_rate
-    for x in range(cfg.epochs):
+    for x in range(checkpoint_epoch + 1, cfg.epochs):
+
+        print("Epoch: %03d" % x)
 
         running_loss = 0.0
         running_correct = 0
@@ -161,29 +140,20 @@ def training(
         # if ((x) % 40 == 0) or (x == cfg.epochs - 1):
         if True:
 
-            model_path = (
-                cfg.data_dir
-                + "/"
-                + arch
-                + "/"
-                + dataset
-                + "/"
-                + arch
-                + "_"
-                + dataset
-                + "_p_"
-                + str(precision)
-                + "_model_"
-                + str(checkpoint_epoch + x)
-                + ".pth"
+            model_path = default_model_path(
+                cfg.data_dir, arch, dataset, precision, fl, ber, pos, x
             )
 
             if not os.path.exists(os.path.dirname(model_path)):
                 os.makedirs(os.path.dirname(model_path))
 
+            if os.path.exists(model_path) and not force:
+                print("Checkpoint already present ('%s')" % model_path)
+                sys.exit(1)
+
             torch.save(
                 {
-                    "epoch": (checkpoint_epoch + x),
+                    "epoch": x,
                     "model_state_dict": model.state_dict(),
                     "optimizer_state_dict": opt.state_dict(),
                     "loss": running_loss / batch_id,
